Google My Business: Extract Location component into its own folder
