+++ client/my-sites/google-my-business/location/location-type.js
